# docs/swagger.yaml
basePath: /api
definitions:
  models.Airport:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      numberOfVotes:
        type: integer
    type: object
  models.Airway:
    properties:
      identifier:
        type: string
      waypoints:
        items:
          $ref: '#/definitions/models.AirwayWaypoint'
        type: array
    type: object
  models.AirwayWaypoint:
    properties:
      airwayId:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/models.Location'
      locationId:
        type: integer
      order:
        type: integer
    type: object
...
  models.CustomFix:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      note:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Location:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      routeSegmentId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Central data API for CTP (Cross the Pond) event management. All /api/*
    endpoints require an X-API-Key header.
  title: CTP API
  version: 1.0.0
paths:
  /airports/{id}:
    delete:
      parameters:
      - description: Airport ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete airport
      tags:
      - airports
    put:
      consumes:
      - application/json
      parameters:
      - description: Airport ID
        in: path
        name: id
        required: true
        type: integer
      - description: Airport fields to update
        in: body
        name: airport
        required: true
        schema:
          $ref: '#/definitions/models.Airport'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Airport'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update airport
      tags:
      - airports
  /airways:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete all airways
      tags:
      - airways
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Airway'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all airways with waypoints
      tags:
      - airways
    post:
      consumes:
      - application/json
      parameters:
      - description: Airway
        in: body
        name: airway
        required: true
        schema:
          $ref: '#/definitions/models.Airway'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Airway'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create airway
      tags:
      - airways
  /airways/{identifier}:
    delete:
      parameters:
      - description: Airway identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete airway by identifier
      tags:
      - airways
  /airways/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Airways
        in: body
        name: airways
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Airway'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Airway'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Bulk create airways
      tags:
      - airways
  /custom-fixes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CustomFix'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all custom fixes
      tags:
      - custom-fixes
    post:
      consumes:
      - application/json
      parameters:
      - description: Custom fix
        in: body
        name: fix
        required: true
        schema:
          $ref: '#/definitions/models.CustomFix'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CustomFix'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create or update custom fix (upsert by identifier)
      tags:
      - custom-fixes
  /custom-fixes/{identifier}:
    delete:
      parameters:
      - description: Custom fix identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete custom fix
      tags:
      - custom-fixes
...
  /events/{eventId}/airports:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Airport'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List airports for an event
      tags:
      - airports
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - description: Airport
        in: body
        name: airport
        required: true
        schema:
          $ref: '#/definitions/models.Airport'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Airport'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create airport
      tags:
      - airports
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"